feat(gradient-border): trigger animation on corresponding empty color selector and block result
Previously, if any or both the color selectors in Gradient Border generator are empty, then the
results are displayed according to the default values of that color selector (red for color1 and
blue for color2). With this fix, the result is blocked and whichever selector is empty is getting an
animated alert.

fix #288